0.3.0 Notification channel added Logs channel added editing owner list was enabled Bugs with channels were fixed

+++ src/Commands/Guild/finish.ts
@@ -1,4 +1,4 @@
-import { VoiceChannel } from 'discord.js';
+import { TextChannel, VoiceChannel } from 'discord.js';
 import {Command} from '../../Interfaces';
 
 export const command: Command = {
@@ -6,7 +6,7 @@ export const command: Command = {
     aliases: ['fg','finish'],
     run: async(client, message, args) => {
         const guild = message.guild;
-        if(guild == null) return;
+        if(!guild) return;
         try {
             const server = await client.database.getServerRelations(guild.id);
             const occasion = server.events.find(event => event.host == message.author.id);
@@ -14,14 +14,22 @@ export const command: Command = {
             if(args.length < 1) throw Error("Event results must be provided. Ask moderation about respond format.");
             // Log ended event
             const voice = guild.channels.cache.get(occasion.voiceChannel);
+            const text = guild.channels.cache.get(occasion.textChannel);
+            if(!text) throw Error("Text channel has been removed, personal statistic will not be updated.");
             if(!voice) throw Error("Voice channel has been removed, personal statistic will not be updated.");
             console.log(`Voice channel is ${voice.name}`);
             await client.ratingController.updateMembers(client, voice as VoiceChannel);
             await client.database.removeOccasion(server.guild, occasion.voiceChannel);
-            await message.channel.send(client.embeds.finishedOccasion, client.embeds.HostCommend(`likeHost.${occasion.host}`, `dislikeHost.${occasion.host}`));
+            await (text as TextChannel).send(client.embeds.finishedOccasion, client.embeds.HostCommend(`likeHost.${occasion.host}`, `dislikeHost.${occasion.host}`));
             setTimeout(() => client.room.delete(guild, occasion.voiceChannel, occasion.textChannel), 10000);
+            //logging 
+            if(server.settings.logging_channel) {
+                const channel = guild.channels.cache.get(server.settings.logging_channel);
+                if(!channel || !channel.isText) return;
+                (channel as TextChannel).send(client.embeds.occasionFinished(args.join(' '), message.author.username, voice.members.size));
+            }
         } catch(error) {
-            await message.channel.send(client.embeds.errorInformation(error));
+            message.channel.send(client.embeds.errorInformation(error));
         }
     }
 };
