+++ src/Events/message.ts
@@ -4,16 +4,15 @@ import {Command, Event} from '../Interfaces';
 export const event: Event = {
     name: 'message',
     run: (client, message: Message) => {
-        if(message.author.bot || !message.guild || !message.content.startsWith(client.config.prefix))
+        if(message.author.bot || !message.content.startsWith(client.config.prefix))
             return;
         const args = message.content.slice(client.config.prefix.length).trim().split(/ +/g);
-        
         const cmd = args.shift()?.toLowerCase();
         if(!cmd) return;
         const command = client.commands.get(cmd) || client.aliases.get(cmd);
         if(command) {
             (command as Command).run(client, message, args);
-            message.delete();
+            if(message.guild) message.delete().catch(error => console.log(error));
         }
     }
 }
